Invert dependencies between data and domain layers
