+++ composeApp/src/desktopMain/kotlin/main.kt
@@ -1,5 +1,7 @@
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
+import com.sildian.apps.restofinder.datalayer.dataLayerModule
+import com.sildian.apps.restofinder.domainlayer.domainLayerModule
 import com.sildian.apps.restofinder.uilayer.MainEntryPoint
 import com.sildian.apps.restofinder.uilayer.uiLayerModule
 import org.koin.core.context.startKoin
@@ -7,7 +9,11 @@ import org.koin.core.context.startKoin
 fun main() = application {
     startKoin {
         printLogger()
-        modules(uiLayerModule)
+        modules(
+            dataLayerModule,
+            domainLayerModule,
+            uiLayerModule,
+        )
     }
     Window(
         onCloseRequest = ::exitApplication,
