+++ dataLayer/build.gradle.kts
@@ -32,6 +32,7 @@ kotlin {
             implementation(project.dependencies.platform(libs.koin.bom))
             implementation(libs.koin.core)
             implementation(projects.core)
+            implementation(projects.domainLayer)
         }
         commonTest.dependencies {
             implementation(libs.kotlin.test)
