+++ dataLayer/src/commonMain/kotlin/com/sildian/apps/restofinder/datalayer/Module.kt
@@ -1,5 +1,6 @@
 package com.sildian.apps.restofinder.datalayer
 
+import com.sildian.apps.restofinder.domainlayer.RestaurantsRepository
 import org.koin.dsl.module
 
 val dataLayerModule = module {
