--- dataLayer/src/commonMain/kotlin/com/sildian/apps/restofinder/datalayer/Restaurant.kt
@@ -1,5 +1,0 @@
-package com.sildian.apps.restofinder.datalayer
-
-data class Restaurant(
-    val name: String,
-)
