+++ domainLayer/src/commonMain/kotlin/com/sildian/apps/restofinder/domainlayer/Module.kt
@@ -1,9 +1,7 @@
 package com.sildian.apps.restofinder.domainlayer
 
-import com.sildian.apps.restofinder.datalayer.dataLayerModule
 import org.koin.dsl.module
 
 val domainLayerModule = module {
-    includes(dataLayerModule)
     factory<GetRestaurantsUseCase> { GetRestaurantsUseCaseImpl(restaurantsRepository = get()) }
 }
