+++ domainLayer/src/commonMain/kotlin/com/sildian/apps/restofinder/domainlayer/Restaurant.kt
@@ -1,5 +1,5 @@
 package com.sildian.apps.restofinder.domainlayer
 
-data class RestaurantDomain(
+data class Restaurant(
     val name: String,
 )
