+++ domainLayer/src/commonMain/kotlin/com/sildian/apps/restofinder/domainlayer/RestaurantsRepository.kt
@@ -1,0 +1,5 @@
+package com.sildian.apps.restofinder.domainlayer
+
+interface RestaurantsRepository {
+    fun getRestaurants(): List<Restaurant>
+}
